Added a new gamemode (Elimination chess)
Engine changes:
Capture sorting (35 elo)
Check extensions (70 elo)
Move generation speedups (95 elo)
Transposition table (30 elo)
Total approx 230 elo

+++ ulci/src/lib.rs
@@ -191,7 +191,7 @@ pub enum Score {
   /// Side to move loses on this move
   Loss(u16),
   /// Side to move has this advantage in centipawns
-  Centipawn(i64),
+  Centipawn(i32),
 }
 
 impl PartialOrd for Score {
